+++ core/specfem/medium_container/impl/domain_container.hpp
