+++ include/medium/dim2/elastic/isotropic_cosserat/domain_properties.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "medium/impl/data_container.hpp"
+#include "specfem/medium_container/impl/domain_container.hpp"
 #include "specfem/point.hpp"
 #include <Kokkos_SIMD.hpp>
